@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   content: ["./src/**/*.{ts,tsx}"],
   theme: {
     extend: {},
